renamed things, updated the imports, added ml-100k dataset using the data loading functionality, changed order of the tests, all tests pass
